(QENG-3267) Rename Beaker::Host::Generator to BeakerHostGenerator
Previously, we were using the Beaker::Host::Generator namespace, where
Beaker::Host is a module, but beaker defines Beaker::Host to be a
class, so you could not require both beaker and beaker-hostgenerator.

This commit renames the module namespace to BeakerHostGenerator so
that it cannot collide with Beaker.

This is a backwards incompatible change if anyone is programmatically
calling Beaker::Host::Generator, but I don't think anyone is.