(QENG-3267) Allow CLI to be called programmatically
Previously, the CLI class relied on the global ARGV which prevented
the CLI from being called programmatically.

This commit changes the initialize method to take an optional array of
arguments. If none are provided, we use the global ARGV as we did
before, although we make a duplicate of it, since we later mutate it,
e.g. pushing "--help".