(QENG-3370) Use BeakerHostGenerator::CLI directly.
This may be necessary to ensure coverage checks can work correctly. Also, it
makes sense to test at this since I know of at least one project that is doing
something similar.

Not so sure about the way this is handling stderr though--just trying to avoid
polluting the minitest console output.