(maint) Decouple hypervisors from generator
This commit is a large refactor that teases apart the generate code from
the hypervisor code.

Previously the VMPooler hypervisor class was a subclass of the Generator
class, which was fine when there was only one hypervisor, but made it
difficult to add support for per-host/multiple hypervisors.

Now, there's a 1-to-many relationship between the Generator
(BeakerHostGenerator::Generator) and the
Hypervisors (BeakerHostGenerator::Hypervisor).

The giant map of OSINFO data that was previously in the
'data/vmpooler.rb' has been promoted and moved into data.rb as the
`osinfo` module function. The parts specific to the VMPooler hypervisor
have been nested under submaps keyed by the name of the hypervisor,
`:vmpooler`.

Adding support for a new hypervisor now requires implementing
BeakerHostGenerator::Hypervisor::Interface, and adding any appropriate
platform data as a submap to the OSINFO map in data.rb, keyed under a
symbol named after the new hypervisor (e.g. `:my_hypervisor`).

The module function `BeakerHostGenerator::Data.get_platform_info` can
then be used when implementing the hypervisor to extract the relevant
platform data under the key `:my_hypervisor`.