(QENG-3808) Support uppercase alphanumeric architectures
This commit adds support for uppercase-only alphanumeric architecture
bits in platform names. For example, we can support things like
"POWER", "POWER7", "S390X" instead of just "32" or "64".

Uppercase letters allow for backwards compatibility when roles are
specified, which come after the architecture bit. This implementation
relies on the fact that the role and arbitrary role characters are
only lowercase.

This allows us to distinguish between the architecture "POWER" and the
role "m" when parsing the spec "POWERm", as in "aix71-POWERm".