(MAINT) Regenerate test fixtures and isolate them
Since we sorted the platforms alphabetically, the fixture generation
results in different filenames. Basically every time we modify
Data#platforms() we should regenerate the fixtures so we have all
platforms covered.

This moves the generated fixtures into a new directory called
`test/fixtures/generated/` so it's easier to work with them and not have
the manually created fixtures get in the way.