(QENG-4204) Add AlwaysBeScheduling hypervisor
This commit adds a new hypervisor called 'abs' that will support
generating configurations necessary for working with CI.next and the
AlwaysBeScheduling service.

Currently, the behavior of the abs hypervisor is basically the same as
the vmpooler hypervisor. The only difference is it doesn't fixup the
nodes to support old PE versions.

Having a built-in hypervisor for ABS alleviates the need for users to
manually override the hypervisor key on each node.