(MAINT) Delete test_files directive to keep fixtures out of the gem
This gemspec directive `test_files` will cause the files it references to be
included in the built gem so they can be used during the `gem test` command.

We do not intend to support the ability for users to run `gem test`, and
furthermore this gem subcommand doesn't even exist anymore.