(MAINT) Remove acceptance testing
This deletes everything related to acceptance testing.

We had no tests, and we have no intentions of adding any. We started to lay out
the groundwork for acceptance testing a while ago with the "acceptance/"
directory, `beaker` test dependency, and the `acceptance` Rake task.

This commit deletes that groundwork, as it is out of date and just getting in
the way at this point.  If we decide to add acceptance tests in the future, it
will likely look different than this anyway.