(QENG-4019) Remove warning message about changing defaults
This warning message was added a while ago to inform users of an upcoming
breaking change. We're no longer certain that we'll be making the breaking
change though, so the warning is a bit misleading.

Furthermore, we're now trying to silence this message when beaker-hostgenerator
is used programatically within Beaker.

There will likely be more changes that are outlined in the linked ticket, such
as adding a new CLI switch to toggle between specific & generic platform names.