(maint) Load lib/test in spec_helper for one off testing
The Rakefile loads these paths so that the specs can require files
needed to run the tests. But if you need to run rspec separately, say
rspec -e 'test-case' because you just want quick feed back on a single
spec, Rake's setup is not involved. This patch conditionally loads the
lib/test dir in the spec_helper as well, so they are present even if
Rake isn't involved.