(PE-21479) Drop use of pe_family/pe_upgrade_family
The pe_family and pe_upgrade_family environment variables were only
being used as a means of signaling that we want pe_dir to be
archives/releases. (If pe_version and pe_family were equal, we'd return
archives/releases for pe_dir).  This behavior was changed in the
previous commit to instead base the tarball source on the version string
format.

The only other behavior that passing family had was that if either
verison or family were nil, pe_dir would be nil. So theoretically, if
you only set pe_version to some valid version, you would get a beaker
config with no pe_dir set, and Beaker could then set pe_dir based on
BEAKER_PE_DIR or pe_dist_dir environment variables.

This commit removes family so that we can clean up pipelines which would
otherwise be setting pe_family just for the purpose of avoiding this
behavior.

It is potentially a breaking change if someone was relying on the
absence of pe_family to allow Beaker to set pe_dir as mentioned above.