(BKR-1548) Totally disable getty when using docker
beaker-docker must run containers in privileged mode with an /sbin/init
entrypoint to successfully simulate a full operating sytem. This
combination can cause the container to open virutal consoles on the host
machine. These getty processes regularly consume 100% of CPU; they are
not automatically killed when the container is destroyed, and must be
cleaned up manually.

tty1 getty services were already disabled as part of the default set of
image commands for docker platforms, but this did not prevent the
behavior for other ttys. Replacing the getty executable with a copy of
/bin/true successfully disables all of them.