Add a vagrant hypervisor
This also adds some base boxes based from the Vagrant Cloud.

Official builds are used for CentOS and Fedora in a generated way. For
Debian Wheezy, Jessie and Stretch 64 bits we also use the official
builds but manually since there's no automated way to map numbers to
releases. For all other platforms we rely on the generic builds which
uses the same naming scheme as our host specs.

We could use the official Ubuntu images in the same way as Debian, but
they're only built for Virtualbox where the generic ones are built for
other provides as well.