(MAINT) Add locale to vmpooler Jap/French defs
This is an outcome of moving the Windows-2012r2 localised vmpooler
platforms to Windows 2019. When trying to test the new BHG strings
in the ci-job-configs repo using the Puppet Repo Jenkins test harness
it was obvious that there were issues with the parsing of the embedded
"locale" string in the BHG name.

So since BHG already supports additional Beaker variables in the
definition (see the use of the "user" variable for the French platform
definitions to allow Windows use a localised Administrator name) it
makes sense to properly embed the "locale" definition here for the
French and Japanese platforms.

This will allow the removal of the  windows2019_ja-64%7Blocale=ja%7D
type string from the Puppet Agent CI job definitions.