(PE-27608) Master branch PEZ logic
In order for PEZ builds to be located correctly we need to add in some
logic. This PR adds to the Master branch logic by looking at the tarball
string and seeing if it is a PEZ tarball. If so point the pe_dir to the
feature directory, where the PEZ tarballs are put.